Final Server

+++ src/main/java/com/samsung/spring/dao/LandmarkDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LandmarkDao extends JpaRepository<Landmark, Long> {
+public interface LandmarkDao extends JpaRepository<Landmark, Integer> {
 }
